+++ ui/src/models/WorkloadSummary.ts
@@ -1,0 +1,6 @@
+import { StatsType } from "./WorkLoadDetail";
+
+export type WorkloadSummaryProps = {
+  stats: StatsType;
+  workloadName: string;
+};
